refactor(reference-data): split saga grpc_handler.go by operation type
Split the 895-line handler into focused files:
- grpc_handler_lifecycle.go: CreateSagaDraft, UpdateSagaDefinition, ActivateSaga, DeprecateSaga
- grpc_handler_query.go: GetSaga, GetActiveSaga, ListSagas, pagination helpers
- grpc_handler_validation.go: ValidateSaga, ValidateSagaDraft, AnalyzeDeprecationImpact, DescribeHandlers
- grpc_handler.go: struct, constructor, error mapping, proto conversions
