+++ services/reference-data/saga/grpc_handler.go
@@ -1,19 +1,11 @@
 // Package saga provides gRPC handlers for saga definition management.
-//
-//meridian:large-file - known oversized file; split tracked in backlog
 package saga
 
 import (
-	"context"
 	"errors"
 	"log/slog"
 	"os"
-	"sort"
-	"strconv"
-	"strings"
-	"time"
 
-	"github.com/google/uuid"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/types/known/timestamppb"
@@ -64,569 +56,6 @@ func (h *RegistryHandler) WithSchemaRegistry(reg *schema.Registry) *RegistryHand
 func (h *RegistryHandler) WithDerivedSchema(s *schema.Schema) *RegistryHandler {
 	h.schemaRegistry = schema.NewRegistryFromSchema(s)
 	return h
-}
-
-// CreateSagaDraft creates a new saga definition in DRAFT status.
-// If a DryRunValidator is configured, the script is validated before persistence.
-// Invalid scripts are rejected with INVALID_ARGUMENT status.
-func (h *RegistryHandler) CreateSagaDraft(
-	ctx context.Context,
-	req *sagav1.CreateSagaDraftRequest,
-) (*sagav1.CreateSagaDraftResponse, error) {
-	version := int(req.Version)
-	if version < 0 {
-		return nil, status.Errorf(codes.InvalidArgument, "version must be non-negative")
-	}
-	if version == 0 {
-		version = 1
-	}
-
-	def := &Definition{
-		ID:                      uuid.New(),
-		Name:                    req.Name,
-		Version:                 version,
-		Script:                  req.Script,
-		Status:                  StatusDraft,
-		IsSystem:                false,
-		PreconditionsExpression: req.PreconditionsExpression,
-		DisplayName:             req.DisplayName,
-		Description:             req.Description,
-	}
-
-	// Mandatory dry-run validation: reject invalid scripts before persistence
-	if h.dryRunValidator != nil && req.Script != "" {
-		dryRunResult, err := h.dryRunValidator.Validate(ctx, req.Script)
-		if err != nil {
-			h.logger.Error("dry-run validation error during draft creation",
-				"saga_name", req.Name,
-				"error", err)
-			return nil, status.Errorf(codes.Internal, "validation error: %v", err)
-		}
-
-		if !dryRunResult.Success {
-			formatter := &validation.HumanReadableFormatter{}
-			errMsg := formatter.Format(dryRunResult)
-
-			h.logger.Warn("saga script validation failed",
-				"saga_name", req.Name,
-				"version", version,
-				"error_count", len(dryRunResult.Errors))
-
-			return nil, status.Errorf(codes.InvalidArgument,
-				"saga script validation failed:\n%s", errMsg)
-		}
-
-		// Store validation metadata on the definition
-		now := time.Now()
-		complexityScore := calculateComplexityScore(dryRunResult.Metrics.HandlerCallCount)
-		handlerCallCount := dryRunResult.Metrics.HandlerCallCount
-		def.ValidationStatus = "PASSED"
-		def.ComplexityScore = &complexityScore
-		def.HandlerCallCount = &handlerCallCount
-		def.ValidatedAt = &now
-	}
-
-	if err := h.registry.CreateDraft(ctx, def); err != nil {
-		return nil, h.mapDomainError(err, "CreateSagaDraft", req.Name)
-	}
-
-	h.logger.Info("saga draft created",
-		"name", def.Name,
-		"version", def.Version,
-		"id", def.ID)
-
-	// Perform reference validation (non-blocking)
-	var validationResult *sagav1.ValidationResult
-	if h.validator != nil {
-		result, err := h.validator.ValidateDraft(ctx, def.ID, def.Script)
-		if err != nil {
-			h.logger.Warn("reference validation failed during draft creation",
-				"saga_id", def.ID,
-				"error", err)
-		} else {
-			validationResult = h.validationResultToProto(result)
-		}
-	}
-
-	// Re-fetch the saga to get timestamps set by the database
-	created, err := h.registry.GetByID(ctx, def.ID)
-	if err != nil {
-		return nil, h.mapDomainError(err, "CreateSagaDraft", req.Name)
-	}
-
-	return &sagav1.CreateSagaDraftResponse{
-		Saga:       h.definitionToProto(created),
-		Validation: validationResult,
-	}, nil
-}
-
-// UpdateSagaDefinition modifies a DRAFT saga definition.
-// Only fields that are explicitly set in the request are updated.
-func (h *RegistryHandler) UpdateSagaDefinition(
-	ctx context.Context,
-	req *sagav1.UpdateSagaDefinitionRequest,
-) (*sagav1.UpdateSagaDefinitionResponse, error) {
-	id, err := uuid.Parse(req.Id)
-	if err != nil {
-		return nil, status.Errorf(codes.InvalidArgument, "invalid saga id: %v", err)
-	}
-
-	// Only update fields that are explicitly provided (proto3 optional fields)
-	updates := &Definition{}
-	if req.Script != nil {
-		updates.Script = *req.Script
-	}
-	if req.PreconditionsExpression != nil {
-		updates.PreconditionsExpression = *req.PreconditionsExpression
-	}
-	if req.DisplayName != nil {
-		updates.DisplayName = *req.DisplayName
-	}
-	if req.Description != nil {
-		updates.Description = *req.Description
-	}
-
-	if err := h.registry.UpdateDefinition(ctx, id, updates); err != nil {
-		return nil, h.mapDomainError(err, "UpdateSagaDefinition", id.String())
-	}
-
-	// Retrieve the updated definition
-	updated, err := h.registry.GetByID(ctx, id)
-	if err != nil {
-		return nil, h.mapDomainError(err, "UpdateSagaDefinition", id.String())
-	}
-
-	h.logger.Info("saga definition updated",
-		"id", id,
-		"name", updated.Name)
-
-	return &sagav1.UpdateSagaDefinitionResponse{
-		Saga: h.definitionToProto(updated),
-	}, nil
-}
-
-// ActivateSaga transitions a saga from DRAFT to ACTIVE.
-func (h *RegistryHandler) ActivateSaga(
-	ctx context.Context,
-	req *sagav1.ActivateSagaRequest,
-) (*sagav1.ActivateSagaResponse, error) {
-	id, err := uuid.Parse(req.Id)
-	if err != nil {
-		return nil, status.Errorf(codes.InvalidArgument, "invalid saga id: %v", err)
-	}
-
-	// Perform activation validation if validator is configured
-	var validationResult *sagav1.ValidationResult
-	if h.validator != nil {
-		saga, err := h.registry.GetByID(ctx, id)
-		if err != nil {
-			return nil, h.mapDomainError(err, "ActivateSaga", id.String())
-		}
-
-		// Use ResolvedScript for platform-ref sagas (Script is empty, script comes from platform)
-		script := saga.Script
-		if script == "" {
-			script = saga.ResolvedScript
-		}
-		result, err := h.validator.ValidateActivation(ctx, id, script)
-		if err != nil {
-			return nil, status.Errorf(codes.Internal, "validation error: %v", err)
-		}
-		validationResult = h.validationResultToProto(result)
-
-		// If validation blocked, don't proceed with activation
-		if result.Status == statusBlocked {
-			return nil, status.Errorf(codes.FailedPrecondition, "saga validation failed: %s", result.FormatReport())
-		}
-	}
-
-	if err := h.registry.ActivateSaga(ctx, id); err != nil {
-		return nil, h.mapDomainError(err, "ActivateSaga", id.String())
-	}
-
-	// Retrieve the activated definition
-	activated, err := h.registry.GetByID(ctx, id)
-	if err != nil {
-		return nil, h.mapDomainError(err, "ActivateSaga", id.String())
-	}
-
-	h.logger.Info("saga activated",
-		"id", id,
-		"name", activated.Name,
-		"version", activated.Version)
-
-	return &sagav1.ActivateSagaResponse{
-		Saga:       h.definitionToProto(activated),
-		Validation: validationResult,
-	}, nil
-}
-
-// DeprecateSaga transitions a saga from ACTIVE to DEPRECATED.
-func (h *RegistryHandler) DeprecateSaga(
-	ctx context.Context,
-	req *sagav1.DeprecateSagaRequest,
-) (*sagav1.DeprecateSagaResponse, error) {
-	id, err := uuid.Parse(req.Id)
-	if err != nil {
-		return nil, status.Errorf(codes.InvalidArgument, "invalid saga id: %v", err)
-	}
-
-	var successorID *uuid.UUID
-	if req.SuccessorId != "" {
-		parsed, err := uuid.Parse(req.SuccessorId)
-		if err != nil {
-			return nil, status.Errorf(codes.InvalidArgument, "invalid successor_id: %v", err)
-		}
-		successorID = &parsed
-	}
-
-	if err := h.registry.DeprecateSaga(ctx, id, successorID); err != nil {
-		return nil, h.mapDomainError(err, "DeprecateSaga", id.String())
-	}
-
-	// Retrieve the deprecated definition
-	deprecated, err := h.registry.GetByID(ctx, id)
-	if err != nil {
-		return nil, h.mapDomainError(err, "DeprecateSaga", id.String())
-	}
-
-	h.logger.Info("saga deprecated",
-		"id", id,
-		"name", deprecated.Name,
-		"version", deprecated.Version,
-		"successorId", req.SuccessorId)
-
-	return &sagav1.DeprecateSagaResponse{
-		Saga: h.definitionToProto(deprecated),
-	}, nil
-}
-
-// GetSaga retrieves a specific saga by ID or by name+version.
-func (h *RegistryHandler) GetSaga(
-	ctx context.Context,
-	req *sagav1.GetSagaRequest,
-) (*sagav1.GetSagaResponse, error) {
-	var def *Definition
-	var err error
-
-	if req.Id != "" {
-		id, parseErr := uuid.Parse(req.Id)
-		if parseErr != nil {
-			return nil, status.Errorf(codes.InvalidArgument, "invalid saga id: %v", parseErr)
-		}
-		def, err = h.registry.GetByID(ctx, id)
-	} else if req.Name != "" {
-		if req.Version < 0 {
-			return nil, status.Errorf(codes.InvalidArgument, "version must be >= 0")
-		}
-		if req.Version == 0 {
-			def, err = h.registry.GetActive(ctx, req.Name)
-		} else {
-			def, err = h.registry.GetDefinition(ctx, req.Name, int(req.Version))
-		}
-	} else {
-		return nil, status.Errorf(codes.InvalidArgument, "either id or name must be provided")
-	}
-
-	if err != nil {
-		identifier := req.Id
-		if identifier == "" {
-			identifier = req.Name
-		}
-		return nil, h.mapDomainError(err, "GetSaga", identifier)
-	}
-
-	return &sagav1.GetSagaResponse{
-		Saga: h.definitionToProto(def),
-	}, nil
-}
-
-// GetActiveSaga retrieves the active saga for a name using tenant resolution.
-func (h *RegistryHandler) GetActiveSaga(
-	ctx context.Context,
-	req *sagav1.GetActiveSagaRequest,
-) (*sagav1.GetActiveSagaResponse, error) {
-	def, err := h.registry.GetActive(ctx, req.Name)
-	if err != nil {
-		return nil, h.mapDomainError(err, "GetActiveSaga", req.Name)
-	}
-
-	return &sagav1.GetActiveSagaResponse{
-		Saga:             h.definitionToProto(def),
-		IsTenantOverride: !def.IsSystem,
-	}, nil
-}
-
-// ListSagas retrieves sagas matching the filter criteria.
-func (h *RegistryHandler) ListSagas(
-	ctx context.Context,
-	req *sagav1.ListSagasRequest,
-) (*sagav1.ListSagasResponse, error) {
-	defs, err := h.fetchSagas(ctx, req.StatusFilter)
-	if err != nil {
-		return nil, status.Errorf(codes.Internal, "failed to list sagas: %v", err)
-	}
-
-	// Filter out system sagas if requested
-	if req.ExcludeSystem {
-		defs = h.filterNonSystemSagas(defs)
-	}
-
-	// Sort for stable pagination
-	h.sortDefinitions(defs)
-
-	// Parse and validate pagination
-	offset, err := h.parsePageToken(req.PageToken)
-	if err != nil {
-		return nil, err
-	}
-
-	// Apply pagination
-	pageSize := h.normalizePageSize(int(req.PageSize))
-	paginatedDefs, nextToken := h.paginate(defs, offset, pageSize)
-
-	// Convert to proto
-	sagas := make([]*sagav1.SagaDefinition, len(paginatedDefs))
-	for i, def := range paginatedDefs {
-		sagas[i] = h.definitionToProto(def)
-	}
-
-	return &sagav1.ListSagasResponse{
-		Sagas:         sagas,
-		NextPageToken: nextToken,
-	}, nil
-}
-
-// fetchSagas retrieves sagas filtered by status.
-func (h *RegistryHandler) fetchSagas(ctx context.Context, statusFilter sagav1.SagaStatus) ([]*Definition, error) {
-	if statusFilter != sagav1.SagaStatus_SAGA_STATUS_UNSPECIFIED {
-		domainStatus := h.protoStatusToDomain(statusFilter)
-		return h.registry.ListByStatus(ctx, domainStatus)
-	}
-
-	// Get all sagas by listing each status
-	drafts, err1 := h.registry.ListByStatus(ctx, StatusDraft)
-	actives, err2 := h.registry.ListByStatus(ctx, StatusActive)
-	deprecated, err3 := h.registry.ListByStatus(ctx, StatusDeprecated)
-
-	if err1 != nil {
-		return nil, err1
-	}
-	if err2 != nil {
-		return nil, err2
-	}
-	if err3 != nil {
-		return nil, err3
-	}
-
-	defs := make([]*Definition, 0, len(drafts)+len(actives)+len(deprecated))
-	defs = append(defs, drafts...)
-	defs = append(defs, actives...)
-	defs = append(defs, deprecated...)
-	return defs, nil
-}
-
-// filterNonSystemSagas returns only non-system sagas.
-func (h *RegistryHandler) filterNonSystemSagas(defs []*Definition) []*Definition {
-	filtered := make([]*Definition, 0, len(defs))
-	for _, def := range defs {
-		if !def.IsSystem {
-			filtered = append(filtered, def)
-		}
-	}
-	return filtered
-}
-
-// sortDefinitions sorts definitions by name, then version for stable pagination.
-func (h *RegistryHandler) sortDefinitions(defs []*Definition) {
-	sort.Slice(defs, func(i, j int) bool {
-		if defs[i].Name == defs[j].Name {
-			return defs[i].Version < defs[j].Version
-		}
-		return defs[i].Name < defs[j].Name
-	})
-}
-
-// parsePageToken parses the offset from a page token.
-func (h *RegistryHandler) parsePageToken(token string) (int, error) {
-	if token == "" {
-		return 0, nil
-	}
-	offset, err := strconv.Atoi(token)
-	if err != nil || offset < 0 {
-		return 0, status.Error(codes.InvalidArgument, "invalid page_token")
-	}
-	return offset, nil
-}
-
-// normalizePageSize ensures page size is within valid bounds.
-func (h *RegistryHandler) normalizePageSize(pageSize int) int {
-	if pageSize <= 0 {
-		return 50
-	}
-	if pageSize > 100 {
-		return 100
-	}
-	return pageSize
-}
-
-// paginate returns a slice of definitions and the next page token.
-func (h *RegistryHandler) paginate(defs []*Definition, offset, pageSize int) ([]*Definition, string) {
-	if offset > len(defs) {
-		offset = len(defs)
-	}
-
-	end := offset + pageSize
-	if end > len(defs) {
-		end = len(defs)
-	}
-
-	nextToken := ""
-	if end < len(defs) {
-		nextToken = strconv.Itoa(end)
-	}
-
-	return defs[offset:end], nextToken
-}
-
-// ValidateSaga validates a saga script using dry-run execution.
-// This validates script syntax, handler existence, type correctness, and runtime behavior.
-func (h *RegistryHandler) ValidateSaga(
-	ctx context.Context,
-	req *sagav1.ValidateSagaRequest,
-) (*sagav1.ValidateSagaResponse, error) {
-	// Return error if validator not configured
-	if h.dryRunValidator == nil {
-		return nil, status.Error(codes.FailedPrecondition, "dry-run validator not configured")
-	}
-
-	// Validate the script using DryRunValidator
-	result, err := h.dryRunValidator.Validate(ctx, req.Script)
-	if err != nil {
-		h.logger.Error("validation failed",
-			"saga_name", req.SagaName,
-			"version", req.Version,
-			"error", err)
-		return nil, status.Errorf(codes.Internal, "validation error: %v", err)
-	}
-
-	// Record validation metrics
-	validation.RecordValidation(req.SagaName, result)
-
-	// Log validation attempt
-	h.logger.Info("saga script validated",
-		"saga_name", req.SagaName,
-		"version", req.Version,
-		"success", result.Success,
-		"error_count", len(result.Errors),
-		"handler_count", result.Metrics.HandlerCallCount)
-
-	// Convert ValidationResult to protobuf response
-	response := &sagav1.ValidateSagaResponse{
-		Success: result.Success,
-		Errors:  make([]*sagav1.ValidationError, 0, len(result.Errors)),
-		Metrics: &sagav1.ComplexityMetrics{
-			HandlerCallCount:    int32(result.Metrics.HandlerCallCount),
-			OperationCount:      int32(result.Metrics.OperationCount),
-			EstimatedDurationMs: int32(result.Metrics.EstimatedDuration.Milliseconds()),
-			ComplexityScore:     int32(calculateComplexityScore(result.Metrics.HandlerCallCount)),
-		},
-	}
-
-	// Convert errors
-	for _, err := range result.Errors {
-		response.Errors = append(response.Errors, &sagav1.ValidationError{
-			Line:       int32(err.Line),
-			Column:     int32(err.Column),
-			Message:    err.Message,
-			Category:   h.errorCategoryToProto(err.Category),
-			Suggestion: "", // Could enhance with handler suggestions later
-		})
-	}
-
-	// Generate formatted report using HumanReadableFormatter
-	formatter := &validation.HumanReadableFormatter{
-		AvailableHandlers: []string{}, // Could inject from schema registry
-	}
-	response.FormattedReport = formatter.Format(result)
-
-	return response, nil
-}
-
-// ValidateSagaDraft validates a saga script without activating it.
-func (h *RegistryHandler) ValidateSagaDraft(
-	ctx context.Context,
-	req *sagav1.ValidateSagaDraftRequest,
-) (*sagav1.ValidateSagaDraftResponse, error) {
-	id, err := uuid.Parse(req.Id)
-	if err != nil {
-		return nil, status.Errorf(codes.InvalidArgument, "invalid saga id: %v", err)
-	}
-
-	// Get the saga
-	saga, err := h.registry.GetByID(ctx, id)
-	if err != nil {
-		return nil, h.mapDomainError(err, "ValidateSagaDraft", id.String())
-	}
-
-	// Validate
-	if h.validator == nil {
-		return &sagav1.ValidateSagaDraftResponse{
-			Validation: &sagav1.ValidationResult{
-				Status: "READY",
-			},
-			Report: "No validator configured",
-		}, nil
-	}
-
-	// Use ResolvedScript for platform-ref sagas (Script is empty, script comes from platform)
-	validationScript := saga.Script
-	if validationScript == "" {
-		validationScript = saga.ResolvedScript
-	}
-	result, err := h.validator.ValidateDraft(ctx, id, validationScript)
-	if err != nil {
-		return nil, status.Errorf(codes.Internal, "validation error: %v", err)
-	}
-
-	return &sagav1.ValidateSagaDraftResponse{
-		Validation: h.validationResultToProto(result),
-		Report:     result.FormatReport(),
-	}, nil
-}
-
-// AnalyzeDeprecationImpact finds all sagas that reference a given instrument.
-func (h *RegistryHandler) AnalyzeDeprecationImpact(
-	ctx context.Context,
-	req *sagav1.AnalyzeDeprecationImpactRequest,
-) (*sagav1.AnalyzeDeprecationImpactResponse, error) {
-	if h.validator == nil {
-		return &sagav1.AnalyzeDeprecationImpactResponse{
-			Dependencies: nil,
-			TotalCount:   0,
-		}, nil
-	}
-
-	deps, err := h.validator.DeprecationImpactAnalysis(ctx, req.InstrumentCode)
-	if err != nil {
-		return nil, status.Errorf(codes.Internal, "failed to analyze deprecation impact: %v", err)
-	}
-
-	protoDeps := make([]*sagav1.SagaDependency, len(deps))
-	for i, dep := range deps {
-		protoDeps[i] = &sagav1.SagaDependency{
-			SagaId:      dep.SagaID.String(),
-			SagaName:    dep.SagaName,
-			SagaVersion: int32(dep.SagaVersion),
-			SagaStatus:  h.domainStatusToProto(dep.SagaStatus),
-			LineNumber:  int32(dep.LineNumber),
-		}
-	}
-
-	return &sagav1.AnalyzeDeprecationImpactResponse{
-		Dependencies: protoDeps,
-		TotalCount:   int32(len(deps)),
-	}, nil
 }
 
 // mapDomainError converts domain errors to appropriate gRPC status codes.
@@ -815,81 +244,4 @@ func calculateComplexityScore(handlerCallCount int) int {
 		return 10
 	}
 	return score
-}
-
-// DescribeHandlers returns the platform handler schema registry grouped by service.
-// Used by the Starlark editor to display available handlers and their parameter types.
-func (h *RegistryHandler) DescribeHandlers(
-	_ context.Context,
-	_ *sagav1.DescribeHandlersRequest,
-) (*sagav1.DescribeHandlersResponse, error) {
-	reg := h.schemaRegistry
-	if reg == nil {
-		h.logger.Warn("DescribeHandlers called without schema registry; returning empty handler list. " +
-			"Use WithSchemaRegistry or WithDerivedSchema to populate handlers.")
-		reg = schema.NewRegistry()
-	}
-
-	// Group handlers by service name (first component of "service.handler" name)
-	handlerNames := reg.ListHandlers()
-	serviceMap := make(map[string][]*sagav1.HandlerSchema)
-	serviceOrder := make([]string, 0)
-
-	for _, fullName := range handlerNames {
-		parts := strings.SplitN(fullName, ".", 2)
-		if len(parts) < 2 {
-			continue
-		}
-		serviceName := parts[0]
-		handlerName := parts[1]
-
-		def, err := reg.GetHandler(fullName)
-		if err != nil {
-			continue
-		}
-
-		params := make([]*sagav1.HandlerParameter, 0, len(def.Params))
-		// Sort parameter names for deterministic output
-		paramNames := make([]string, 0, len(def.Params))
-		for paramName := range def.Params {
-			paramNames = append(paramNames, paramName)
-		}
-		sort.Strings(paramNames)
-
-		for _, paramName := range paramNames {
-			paramDef := def.Params[paramName]
-			params = append(params, &sagav1.HandlerParameter{
-				Name:        paramName,
-				Type:        string(paramDef.Type),
-				Required:    paramDef.Required,
-				EnumValues:  paramDef.Values,
-				Description: paramDef.Description,
-			})
-		}
-
-		if _, exists := serviceMap[serviceName]; !exists {
-			serviceOrder = append(serviceOrder, serviceName)
-		}
-		serviceMap[serviceName] = append(serviceMap[serviceName], &sagav1.HandlerSchema{
-			Name:              handlerName,
-			Description:       def.Description,
-			Parameters:        params,
-			IsExternal:        def.External,
-			CompensateHandler: def.Compensate,
-		})
-	}
-
-	sort.Strings(serviceOrder)
-
-	services := make([]*sagav1.HandlerServiceSchema, 0, len(serviceOrder))
-	for _, serviceName := range serviceOrder {
-		services = append(services, &sagav1.HandlerServiceSchema{
-			Name:     serviceName,
-			Handlers: serviceMap[serviceName],
-		})
-	}
-
-	return &sagav1.DescribeHandlersResponse{
-		Services: services,
-	}, nil
 }
